part4-Done

+++ part4/blog/controllers/blogRouter.js
@@ -1,24 +1,43 @@
 const blogRouter=require("express").Router()
-const note = require("../../notes_backend/models/note")
 const Blog=require("../models/blog")
+const User=require("../models/user")
+const jwt=require("jsonwebtoken")
+
 
 blogRouter.get('/', async(request, response) => {
-  const res=await Blog.find({})
+  const res=await Blog.find({}).populate("user")
   response.json(res)
 })
 
 blogRouter.post('/', async(request, response) => {
+  const body=request.body
+  
   if(request.body.likes===undefined)request.body.likes=0;
   if(request.body.url===undefined)return response.status(400).end();
   if(request.body.title===undefined)return response.status(400).end();
 
-  const blog = new Blog(request.body)
+  let userFound = request.user
+  if(!userFound){
+    return response.status(400).json({error:"userid not found"})
+  }
+  body.user = userFound._id
 
+  const blog = new Blog(request.body)
   const result=await blog.save()
+
+  userFound.blogs=userFound.blogs.concat(result._id)
+  await userFound.save()
   response.status(201).json(result)
 })
 
 blogRouter.delete("/:id",async(request,response)=>{
+  const decodedToken=request.user
+  
+  const blog=await Blog.findById(request.params.id)
+
+  if(blog.user.toString()!=decodedToken.id){
+    return response.status(401).json({error:"deletion not possible, different IDS"})
+  }
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
